requirements.txt
absl-py==0.15.0
adblockparser==0.7
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
asgiref==3.3.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.2.0
aws-deepracer-control-v2==0.1.5
awsdeepracer-control==0.1.3
base58==2.1.1
beautifulsoup4==4.9.3
bitarray==1.2.2
black==21.11b1
bs4==0.0.1
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.1.1
certifi==2021.10.8
cffi==1.14.2
chardet==3.0.4
charset-normalizer==2.0.9
cheroot==8.5.2
CherryPy==18.6.0
cherrypy-cors==1.6
clang==5.0
click==8.0.3
coinbase==2.1.0
cryptography==3.4.6
cycler==0.10.0
Cython==3.0.0a11
cytoolz==0.11.2
dataclassy==0.11.1
DBN==0.1.0
decorator==4.4.2
deepracer-vehicle-api==0.0.10
Deprecated==1.2.13
distlib==0.3.1
Django==3.1.6
dnspython==2.1.0
docopt==0.6.2
eip712==0.1.0
eth-abi==2.1.1
eth-account==0.5.6
eth-brownie==1.17.2
eth-event==1.2.3
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
execnet==1.9.0
ExifRead==2.3.2
filelock==3.0.12
firebase-admin==4.3.0
Flask==1.1.2
flatbuffers==1.12
frozenlist==1.2.0
future==0.18.2
gast==0.4.0
geojson==2.5.0
glfw==2.3.0
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.6
google-cloud-core==1.4.1
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.0
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
graphviz==0.19.1
grpcio==1.41.1
h5py==3.1.0
hexbytes==0.2.2
httpagentparser==1.9.1
httplib2==0.18.1
hypothesis==6.27.3
idna==3.3
imbalanced-learn==0.8.1
imblearn==0.0
importlib-metadata==4.8.2
inflection==0.5.0
iniconfig==1.1.1
inputs==0.5
ipaddr==2.2.0
ipfshttpclient==0.8.0a2
ipwhois==1.0.0
itsdangerous==1.1.0
jaraco.classes==3.2.1
jaraco.collections==3.2.0
jaraco.functools==3.2.1
jaraco.text==3.5.0
Jinja2==2.11.2
joblib==1.0.1
jsonschema==3.2.0
jwcrypto==1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
lru-dict==1.1.7
lxml==4.6.2
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.3
mne==0.24.1
more-itertools==8.7.0
msgpack==1.0.0
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
networkx==2.5
numpy==1.19.2
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
osmapi==3.0.0
OSMPythonTools==0.3.5
packaging==21.3
pandas==1.3.2
parsimonious==0.8.1
pathspec==0.9.0
phonenumbers==8.12.19
Pillow==8.1.2
pipreqs==0.4.11
platformdirs==2.4.0
playsound==1.3.0
pluggy==1.0.0
portend==2.7.1
prompt-toolkit==3.0.23
protobuf==3.19.1
psutil==5.8.0
publicsuffixlist==0.7.7
py==1.11.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyBluez==0.23
pycocotools==2.0
pycparser==2.20
pycryptodome==3.12.0
pygame==2.0.3
pygexf==0.2.2
Pygments==2.10.0
pygments-lexer-solidity==0.7.0
PyJWT==1.7.1
pynmea2==1.18.0
pynmeagps==1.0.16
pyobjc-core==7.3
pyobjc-framework-Cocoa==7.3
pyOpenSSL==20.0.1
pyparsing==3.0.6
PyPDF2==1.26.0
pyrsistent==0.18.0
pyserial==3.5
PySocks==1.7.1
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-dateutil==2.8.1
python-decouple==3.5
python-docx==0.8.10
python-dotenv==0.16.0
python-jwt==3.3.0
python-pptx==0.6.18
python-whois==0.7.3
pythx==1.6.1
pytz==2020.1
PyYAML==5.4.1
redis==4.3.4
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.10.1
rlp==2.0.1
rsa==4.6
schedule==1.1.0
scikit-learn==0.24.2
scipy==1.7.1
secure==0.2.1
semantic-version==2.8.5
six==1.16.0
sklearn==0.0
sortedcontainers==2.4.0
soupsieve==2.2
spotipy==2.20.0
sqlparse==0.4.1
stem==1.8.0
tempora==4.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==2.2.0
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
torch==1.10.2
torchvision==0.11.3
torchviz==0.0.2
tqdm==4.62.3
tweepy==4.4.0
twilio==7.5.0
typed-ast==1.5.1
typing-extensions==3.10.0.2
ujson==5.5.0
uritemplate==3.0.1
urllib3==1.26.7
varint==1.0.2
virtualenv==20.4.2
vvm==0.1.0
vyper==0.3.1
wcwidth==0.2.5
web3==5.25.0
websockets==9.1
Werkzeug==1.0.1
wrapt==1.13.3
xarray==0.20.2
XlsxWriter==1.3.7
yarg==0.1.9
yarl==1.7.2
zc.lockfile==2.0
zipp==3.6.0
